AnnA.NonDimensionalize — Method

NonDimensionalize(p::AbstractParameters)

Performs the nondimensionalisation of the input parameters and returns a NodimParameters struct

AnnA._resample — Method

_resample(data::Array{T,2},N::Int,::Val{:xlog}) where T <: Number

resamples the data to a log scaled x-axis by appling peacwise quadratic fit arround the 'N' samle points ( this apporach is comparable to loesssmotthing with non-constant stepp size)

AnnA.calc_ideality — Method

calc_ideality(s::IVSolution;at_V=:all)

returns the ideality of an IV solution at the unitful voltage at_V. If no number is provided a array is returned.

AnnA.init_guess — Method

init_guess(g::Grid,ndim::NodimParameters)

Returns an initial guess for NLsolve root finding. The NodimParameter value is needed to guarantee consitancy of the guess.

AnnA.initial_conditions! — Method

initial_conditions!(c::Cell)

Inplace mutation variant of initial_conditions(c::Cell)`

AnnA.initial_conditions — Method

initial_conditions(c::Cell)

Calculates the initial condition of a given Cell using NLsolve. Returns the steady state solution vector.

AnnA.pulse — Method

pulse(; Δh=1.0, h₀=0.0, w=1.0 - 2.0e-12, Δt=1.0e-12, tₑ=1.0)

creates a pulse. Monotonic interpolation is used, so the pulse shape is differentiable, allowing for autodiff.

Arguments:

Δh: Pulse Amplitude

h₀: Baseline

w: Pulse width

Δt: rise and fall time

tₑ: end time (end of fall)

AnnA.rdim_sol — Method

rdim_sol(sol,p::AbstractParameters)

a single time slice sol of the ODESolution is redimensionalized with p

AnnA.rdim_sol — Method

rdim_sol(sol::DiffEqBase.ODESolution,t)

returns a redimensionalized solution vector at time t, the time is treated as seconds if no unit is given.

AnnA.rdim_sol — Method

rdim_sol(sol::DiffEqBase.ODESolution)

Redimensionalizes the complete ODESolution at each time step.

AnnA.unpac_struct — Method

unpac_struct(s...)

Returns a OrderedDict containing all field => value pairs. Allows for multiple structs as input. The result is then a single Dict. The letter arguments will overwrite previous one is existing already

Base.length — Method

length(g::Grid)

gives the length of the input-array which is 4*g.N+4+2*g.Nₑ+2*g.Nₕ

AnnA.Gen_function — Type

struct Gen_function{T, F, S} <: Function

after constructing this type using g! = Gen_function(αb, dir, light, tion) givs access to the generation function g!(R, n, p)

#Arguments:

αb: Nondim absorption coefficient

dir: Light direction +1/-1

light: scalar tempoal function of generation

tion: ionic timescale, needed for redimensionalize the time

AnnA.Gen_function — Method

(g!::Gen_function)(G, x, t)

Inplace generation function

#Arguments:

G: Buffer Array

x: Nodimensionalized positions Array

t: Nondimensionalized time

AnnA.IVProblem — Method

IVProblem(
    parm::Parameters,
    range::AbstractArray,
    rate::Unitful.AbstractQuantity;
    double_sweep = true,
    alg_control = AlgControl(dtmin = 1e-20,
        dt = 1e-6,
        reltol = 1e-4,
        abstol = 1e-12,
        tend = (maximum(range) - minimum(range)) / abs(rate)
    ),
)

Creates an IVProblem. The voltage parameter V defined in parm gets overwritten by the linear voltage sweep defined as V = t-> first(range) + rate*t. AlgControl.tend is forced to be (maximum(range) - minimum(range)) / abs(rate), an overwrite can be done on the finalized object using Setfield: p = Setfield.setproperties(p::IVProblem, alg_control=AlgControl(...))

AnnA.JscVocProblem — Method

AnnA.LinOperator — Type

struct LinOperator{T, S} <: AbstractOperators

AnnA.NodimParameters — Type

struct NodimParameters{T, F <: Function, Fl <: Function, Fr <: Function, Fg <: Function, Fv <: Function}

holds all nondimensionalized parameters

Arguments:

λ: DESCRIPTION

λ²: DESCRIPTION

δ: DESCRIPTION

χ: DESCRIPTION

σ: DESCRIPTION

κₚ: DESCRIPTION

κₙ: DESCRIPTION

αb: DESCRIPTION

dpt: DESCRIPTION

dpf: DESCRIPTION

Vbi: Vbi optained from bolzmann statistics

wₑ: DESCRIPTION

wₕ: DESCRIPTION

κₑ: DESCRIPTION

κₕ: DESCRIPTION

rₑ: DESCRIPTION

rₕ: DESCRIPTION

λₑ: DESCRIPTION

λₑ²: DESCRIPTION

λₕ²: DESCRIPTION

λₕ: DESCRIPTION

kₑ: proportionality factor between ETL / PVSK electron densities

kₕ: proportionality factor between PVSK / HTL hole densities

R: Bulk recombination function

Rₗ: ETL/PVSK surface recombnation fnction

Rᵣ: PVSK/HTL surface recombination function

G: Carrier generation function

V: Voltage/kt

AnnA.Operators — Method

Operators(g::Grid;dense::Bool=true)

Compute the operators based on a given Grid. Use the dense keyword to switch between sparse ord dens representation

AnnA.Pot_function — Type

struct Pot_function{T, F} <: Function

after constructing this type using v = Pot_function(VT, V, tion) givs access to the generation function v(t)

#Arguments:

VT: thermal voltage

V: temporal potential function

tion: ionic timescale, needed for redimensionalize the time

AnnA.Pot_function — Method

(v::Pot_function)(t)

Returns the nondimensionalized potential at time t

AnnA.Pulse — Type

struct Pulse{A, AA, T, TT, I} <: Function

Holds all information about the pulse formation.

Arguments:

Δh: Pulse Amplitude

h₀: Baseline

w: Pulse width

Δt: rise and fall time

tₑ: end time (end of fall)

ts: array of timepoints

hs: array of amplitude points

itp: interploating function

AnnA.Pulse — Method

(p::Pulse)(t)

calls the interpolating function from Pulse

AnnA.Rec_function — Type

struct Rec_function{T} <: Function

after constructing this type using r = Rec_function(nᵢ², kk, γ, τᵣ, rtrap) givs access to the recombination function r(R, n, p)

#Arguments:

nᵢ²: nondimensionalized intrinsic carrier square

kk: nondimensionalized bimolecular recombination rate

γ: nondimensionalized rate constant for hole-dominated recombination

τᵣ: ratio of SRH carrier lifetime

rtrap: nondimensionalized k₂ (deep trap) parameter

AnnA.Rec_function — Method

(r!::Rec_function)(R::AbstractArray, n::AbstractArray, p::AbstractArray)

Call the inplace recombination function. This is where the bulk recombination is calculated

#Arguments:

R: Buffer Array

n: Electron Array

p: Hole Array

AnnA.Rec_function — Method

(r::Rec_function)(n::T, p::T) where T

Out of place definition of the recombination function. Calculates the recombination for a given n and p. This is intended to calculate the sruface crecombinations

LinearAlgebra.Tridiagonal — Method

(LinearAlgebra.Tridiagonal(N::Integer, a::T, b::T, c::T) where T <: Union{Number, AbstractArray{L, 1}}) where L <: Number

A simple wrapper for creating tridiagonal matrices a bit more convineantly

Arguments:

N: Size of the NxN tridiagonal

a: row a

b: row b

c: row c

Example

julia> AnnA.Tridiagonal(4,1,-2,-1)
4×4 LinearAlgebra.Tridiagonal{Int64, Vector{Int64}}:
 -2  -1   ⋅   ⋅
  1  -2  -1   ⋅
  ⋅   1  -2  -1
  ⋅   ⋅   1  -2