Predefined Protocols

Convenient wrappers are predefined to simplify the simulation of most important protocols.

IV Simulations

The simulation of an IV curve can be done by the IVProblem constructor

AnnA.IVProblem — Type

IVProblem(
    parm::Parameters,
    range::Union{AbstractArray,Tuple},
    rate;
    double_sweep = true,
    alg_control = AlgControl(dtmin = 1e-20,
        dt = 1e-6,
        reltol = 1e-4,
        abstol = 1e-12,
        tend = abs(sum(diff(range))) / abs(rate) * (1+double_sweep)
    ),  
)

Creates an IVProblem. The voltage parameter V defined in parm will be overwritten by the linear voltage sweep defined as V = t-> first(range) + rate*t. AlgControl.tend is forced to be abs(sum(diff(range))) / abs(rate) * (1+double_sweep), an overwrite can be done on the finalized object using Setfield: p = Setfield.setproperties(p::IVProblem, alg_control=AlgControl(...)).

Units

If no units provided for range and rate, V and V/s is assumed.

Here is a simple example on how to use

using AnnA
using Unitful, UnitfulRecipes
parm = Parameters(light = t -> 1.0,
    vₙₕ= 10u"m/s" ,                 # electron surface recombination vel. at HTM
    vₚₕ= 0.01u"m/s" ,               # hole surface recombination vel. at HTM
    N=500,                          # grid size
    N₀=1e18u"cm^-3"                 # ionic concentration
)
prob = IVProblem(parm, [-0.5,1.7]u"V", 0.2u"V/s")
sol  = solve(prob)

AnnA.ProblemSolution{AnnA.IV}(t: [0.0, 2.000763597981684e-7, 8.221450195442589e-7, 3.9415673601305446e-6, 1.3454664849786941e-5, 3.5487547877089394e-5, 7.778842417898983e-5, 0.00014893460880829027, 0.00026185105948919554, 0.000434644247362265  …  0.10239103002829579, 0.10327538257100341, 0.10415973511371103, 0.10504408765641865, 0.10592844019912627, 0.10681279274183389, 0.10769714528454151, 0.10858149782724913, 0.10946585036995675, 0.10995801813963932]
u: [[0.03281147383244288, 0.032953589394029474, 0.03310017678532636, 0.0332513905748598, 0.0334073914275084, 0.03356834638381078, 0.03373442915394231, 0.03390582042721159, 0.034082708197987616, 0.03426528810902501  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.03281147410186844, 0.03295358952072399, 0.033100176838019704, 0.03325139059450508, 0.033407391434517576, 0.033568346386687126, 0.033734429155687216, 0.033905820428718064, 0.034082708199470325, 0.03426528811051934  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.03281147610506772, 0.0329535909942592, 0.03310017788090742, 0.033251391303284035, 0.03340739189776635, 0.03356834667931776, 0.03373442933613575, 0.033905820539360525, 0.03408270826903513, 0.034265288157266684  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.032811497937496784, 0.03295361006184725, 0.03310019437292021, 0.0332514054244775, 0.033407403864538936, 0.03356835671490023, 0.03373443766535893, 0.03390582738348592, 0.03408271384086788, 0.0342652926567926  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.0328116274733052, 0.032953731231364955, 0.033100307334803784, 0.03325151036483662, 0.03340750099694742, 0.03356844628002786, 0.03373451993019455, 0.03390590264013982, 0.03408278240493148, 0.03426535486523525  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.032812141138076976, 0.03295422622251592, 0.03310078361313352, 0.03325196792097104, 0.033407939853253896, 0.03356886649248175, 0.03373492159017198, 0.0339062858761079, 0.034083147384003504, 0.034265701794553856  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.032813677086453504, 0.03295572843942162, 0.033102251680182236, 0.0332534014428601, 0.03340933846033359, 0.03357022984344196, 0.033736249374840425, 0.033907577818352526, 0.03408440324472901, 0.03426692137477964  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.032817430377193146, 0.03295942968460287, 0.03310589994128106, 0.033256995790215536, 0.03341287797563411, 0.0335737136223964, 0.03373967653004034, 0.033910947482332, 0.03408771457322736, 0.034270173550211495  …  0.9999999999999999, 0.9999999999999999, 0.9999999999999999, 0.9999999999999999, 0.9999999999999999, 0.9999999999999999, 1.0, 1.0, 1.0, 1.0], [0.0328256932977001, 0.03296761929314781, 0.033114014668934565, 0.033265034057755764, 0.03342083819556782, 0.033581594201183236, 0.033747475870533186, 0.03391866398644668, 0.034095346644855314, 0.034277719598389066  …  0.9999999999999999, 0.9999999999999999, 0.9999999999999999, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.032842734914592324, 0.032984566308020304, 0.03313086485528611, 0.03328178515715165, 0.03343748791896038, 0.03359814023042649, 0.03376391586010702, 0.0339349955654072, 0.034111567419028696, 0.034293827152827284  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0]  …  [0.2213904212316825, 0.22200924511038295, 0.22264592715838444, 0.22330099652872434, 0.22397499866125223, 0.22466849580344808, 0.22538206754778498, 0.22611631138603644, 0.22687184328095014, 0.22764929825570226  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.2140598706866152, 0.21466834667837817, 0.21529441753042114, 0.2159386065829552, 0.21660145339529827, 0.21728351427075368, 0.2179853627985287, 0.2187075904131311, 0.21945080697170735, 0.22021564134978008  …  0.9999999999999999, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.20685693566270058, 0.2074548946905163, 0.20807017947658127, 0.2087033073287887, 0.20935481169731157, 0.2100252427031436, 0.2107151676841586, 0.2114251717591642, 0.21215585841045403, 0.21290785008535934  …  0.9999999999999999, 0.9999999999999999, 0.9999999999999999, 0.9999999999999999, 0.9999999999999999, 0.9999999999999999, 1.0, 1.0, 1.0, 1.0], [0.19977981992895977, 0.20036708281398444, 0.20097139623824056, 0.20159327124649942, 0.20223323493943043, 0.20289183100538885, 0.20356962027018252, 0.20426718126533128, 0.2049851108153635, 0.20572402464469278  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.19282777632071776, 0.19340415817048703, 0.19399730901724596, 0.19460773340346535, 0.1952359518312003, 0.19588250129669058, 0.19654793584338257, 0.19723282713392348, 0.1979377650417111, 0.19866335826258502  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.18600089341295623, 0.1865662075642344, 0.1871480027482964, 0.18774677676054258, 0.1883630432493929, 0.18899733225324825, 0.18965019075629608, 0.19032218326374803, 0.19101389239713154, 0.19172591951026188  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.17929991284191593, 0.17985397439708825, 0.18042422259815696, 0.18101114824572193, 0.18161525787633037, 0.18223707430132965, 0.1828771371649669, 0.18353600352235924, 0.184214248437994, 0.18491246560542673  …  0.9999999999999999, 0.9999999999999999, 0.9999999999999999, 0.9999999999999999, 0.9999999999999999, 0.9999999999999999, 1.0, 1.0, 1.0, 1.0], [0.1727260724190001, 0.17326870140717585, 0.17382721632012368, 0.1744021007154338, 0.17499385375887092, 0.1756029907646324, 0.1762300437552348, 0.17687556204168656, 0.17754011282464416, 0.17822428181725922  …  1.0, 1.0, 1.0000000000000002, 1.0000000000000002, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.1662809713601031, 0.16681199554892334, 0.1673585987763625, 0.16792125711064282, 0.168500462089559, 0.16909672126164077, 0.1697105587473023, 0.17034251582067034, 0.17099315151282463, 0.17166304323719686  …  0.9999999999999999, 0.9999999999999999, 0.9999999999999999, 0.9999999999999999, 0.9999999999999999, 1.0, 1.0, 1.0, 1.0, 1.0], [0.1627505214689804, 0.1632750163362135, 0.16381491686624902, 0.1643706948539199, 0.1649428374817799, 0.1655318478615483, 0.16613824559572973, 0.16676256736012132, 0.16740536750795884, 0.1680672186964707  …  1.0000000000000002, 1.0000000000000002, 1.0000000000000002, 1.0000000000000002, 1.0000000000000002, 1.0000000000000002, 1.0, 1.0, 1.0, 1.0]], Parameters(500, 8.854187817e-12 F m^-1, 1.6021766209e-19 C, 8.61733035e-5 eV K^-1, 4.0e-7 m, 24.1, -3.7 eV, -5.3 eV, 0 m^-3, 0.00017 m^2 s^-1, 0.00017 m^2 s^-1, 0.2, 0.2, 1.0e18 cm^-3, 6.5e-8 m^2 s^-1, 0.58 eV, false, 300 K, 1.3e7 m^-1, 1.4e21 m^-2 s^-1, 1, Main.__atexample__named__iv.var"#1#2"(), AnnA.var"#73#74"{Vector{Float64}, Float64, Float64, Bool}([-0.5, 1.7], 0.14279966607226333, 2.2, false), 1.0e6 m^2 V A^-1, 0.0001 m^2 V A^-1, 3.0e-7 s, 3.0e-7 s, 3.22e-17 m^3 s^-1, 0 m^4 s^-1, 0 m^4 s^-1, 0 m s^-1, 0 m s^-1, 10 m s^-1, 0.01 m s^-1, 1.0e18 cm^-3, 1.5, -4.0 eV, 1.0e-7 m, 3, 1.0e-7 m^2 s^-1, 1.0e18 cm^-3, 12, -5 eV, 1.0e-7 m, 3, 1.0e-7 m^2 s^-1), 243×22 DataFrame
 Row │ t               V            j                light    nₑ               ⋯
     │ Quantity…       Quantity…    Quantity…        Float64  Array…           ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │  0.0 s          -0.5 V       -223.006 A m^-2      1.0  Unitful.Quantity ⋯
   2 │  4.00305e-5 s   -0.499992 V  -223.006 A m^-2      1.0  Unitful.Quantity
   3 │  0.000164492 s  -0.499967 V  -223.006 A m^-2      1.0  Unitful.Quantity
   4 │  0.000788614 s  -0.499842 V  -223.006 A m^-2      1.0  Unitful.Quantity
   5 │  0.00269196 s   -0.499462 V  -223.006 A m^-2      1.0  Unitful.Quantity ⋯
   6 │  0.00710022 s   -0.49858 V   -223.006 A m^-2      1.0  Unitful.Quantity
   7 │  0.0155636 s    -0.496887 V  -223.005 A m^-2      1.0  Unitful.Quantity
   8 │  0.0297983 s    -0.49404 V   -223.004 A m^-2      1.0  Unitful.Quantity
  ⋮  │       ⋮              ⋮              ⋮            ⋮                      ⋱
 237 │ 21.0168 s       -0.303368 V  -223.058 A m^-2      1.0  Unitful.Quantity ⋯
 238 │ 21.1938 s       -0.338755 V  -223.059 A m^-2      1.0  Unitful.Quantity
 239 │ 21.3707 s       -0.374143 V  -223.059 A m^-2      1.0  Unitful.Quantity
 240 │ 21.5477 s       -0.409531 V  -223.059 A m^-2      1.0  Unitful.Quantity
 241 │ 21.7246 s       -0.444918 V  -223.059 A m^-2      1.0  Unitful.Quantity ⋯
 242 │ 21.9015 s       -0.480306 V  -223.059 A m^-2      1.0  Unitful.Quantity
 243 │ 22.0 s          -0.5 V       -223.06 A m^-2       1.0  Unitful.Quantity
                                                 18 columns and 228 rows omitted, Unitful.Quantity{Float64, 𝐋, Unitful.FreeUnits{(m,), 𝐋, nothing}}[-1.0e-7 m, -9.998853428048007e-8 m, -9.997579123813249e-8 m, -9.996162895659301e-8 m, -9.994588984187134e-8 m, -9.992839891160143e-8 m, -9.990896190260756e-8 m, -9.988736317867561e-8 m, -9.986336341890089e-8 m, -9.983669706540915e-8 m  …  -1.929304923610431e-10 m, -1.6330293459084766e-10 m, -1.3663658109910326e-10 m, -1.1263682132438556e-10 m, -9.10380973924374e-11 m, -7.160108839856516e-11 m, -5.411015812866737e-11 m, -3.8371043406987266e-11 m, -2.4208761867500205e-11 m, -1.1465719519920814e-11 m], Unitful.Quantity{Float64, 𝐋, Unitful.FreeUnits{(m,), 𝐋, nothing}}[0.0 m, 1.1462291596431307e-11 m, 2.3243781715742705e-11 m, 3.535333970741128e-11 m, 4.780008025975224e-11 m, 6.059337012358501e-11 m, 7.374283501682122e-11 m, 8.725836671428233e-11 m, 1.0115013032767627e-10 m, 1.1542857178075127e-10 m  …  3.9988457142821927e-7 m, 3.998988498696723e-7 m, 3.999127416332857e-7 m, 3.9992625716498315e-7 m, 3.999394066298764e-7 m, 3.9995219991974026e-7 m, 3.999646466602926e-7 m, 3.9997675621828425e-7 m, 3.999885377084036e-7 m, 4.0e-7 m], Unitful.Quantity{Float64, 𝐋, Unitful.FreeUnits{(m,), 𝐋, nothing}}[4.0001146571951993e-7 m, 4.000242087618675e-7 m, 4.00038371043407e-7 m, 4.0005411015812863e-7 m, 4.000716010883986e-7 m, 4.000910380973924e-7 m, 4.0011263682132437e-7 m, 4.001366365810991e-7 m, 4.0016330293459087e-7 m, 4.00192930492361e-7 m  …  4.998366970654091e-7 m, 4.998633634189008e-7 m, 4.998873631786755e-7 m, 4.999089619026075e-7 m, 4.999283989116014e-7 m, 4.999458898418713e-7 m, 4.99961628956593e-7 m, 4.999757912381324e-7 m, 4.9998853428048e-7 m, 5.0e-7 m])

The ProblemSolution object contains also grid and spatial information. All timesteps are stored in a DataFrame an can be acessed via the df field of sol. If we just want to plot the IV characteristics we can do:

using Plots
using UnitfulRecipes  # To interface Unitful with Plots

sol=sol.df
sol.j = sol.j .|>u"mA/cm^2" # scale to common units
plt = plot(sol.V[sol.fwd], sol.j[sol.fwd],label="Forward", ylims=(-25,40),xlims=(-0.5,1.3),legend=:topleft);
plot!(plt,sol.V[.!sol.fwd], sol.j[.!sol.fwd],label="Backward");
plt

Jsc vs. Voc Curve

The wrapper implements $j_{sc}(V_{oc})$ simulation by two consecutive simulation runs, where the illumination is increased exponentially over time. This reflects the aspect of a slow $V_{oc}$ built up under low illumination intensities.

Open Circuit Voltage Decay (OCVD)

To simulate a open circuit voltage decay a OCVDProblem is implemented:

prob_ocvd = OCVDProblem(
    parm,       # input parameter set
    50u"s",     # illumination time
    1e5u"s",    # time the decay will be simulated to
)

sol = solve(prob_ocvd)
plot(sol.t_decay,sol.V_decay,xscale=:log10,xlims=(1e-8,1e5))