IV Simulations

The simulation of an IV curve can be done by the IVProblem constructor

AnnA.IVProblem — Type

IVProblem(
    parm::Parameters,
    range::Union{AbstractArray,Tuple},
    rate;
    double_sweep = true,
    alg_control = AlgControl(dtmin = 1e-20,
        dt = 1e-6,
        reltol = 1e-4,
        abstol = 1e-12,
        tend = abs(sum(diff(range))) / abs(rate) * (1+double_sweep)
    ),  
)

Creates an IVProblem. The voltage parameter V defined in parm will be overwritten by the linear voltage sweep defined as V = t-> first(range) + rate*t. AlgControl.tend is forced to be abs(sum(diff(range))) / abs(rate) * (1+double_sweep), an overwrite can be done on the finalized object using Setfield: p = Setfield.setproperties(p::IVProblem, alg_control=AlgControl(...)).

Units

If no units provided for range and rate, V and V/s is assumed.

Here is a simple example on how to us

using AnnA
using Unitful, UnitfulRecipes
parm = Parameters(light = t -> 1.0,
    vₙₕ= 10u"m/s" ,                 # electron surface recombination vel. at HTM
    vₚₕ= 0.01u"m/s" ,               # hole surface recombination vel. at HTM
    N=500,                          # grid size
    N₀=1e18u"cm^-3"                 # ionic concentration
)
prob = IVProblem(parm, [-0.5,1.7]u"V", 0.2u"V/s")
sol  = solve(prob)

AnnA.ProblemSolution(Parameters(500, 8.854187817e-12 F m^-1, 1.6021766209e-19 C, 8.61733035e-5 eV K^-1, 4.0e-7 m, 24.1, -3.7 eV, -5.3 eV, 0 m^-3, 0.00017 m^2 s^-1, 0.00017 m^2 s^-1, 0.2, 0.2, 1.0e18 cm^-3, 6.5e-8 m^2 s^-1, 0.58 eV, false, 300 K, 1.3e7 m^-1, 1.4e21 m^-2 s^-1, 1, Main.ex-iv.var"#1#2"(), AnnA.var"#26#27"{Vector{Float64}, Float64, Float64, Bool}([-0.5, 1.7], 0.14279966607226333, 2.2, false), 1.0e6 m^2 V A^-1, 3.0e-7 s, 3.0e-7 s, 3.22e-17 m^3 s^-1, 0 m^4 s^-1, 0 m^4 s^-1, 0 m s^-1, 0 m s^-1, 10 m s^-1, 0.01 m s^-1, 1.0e18 cm^-3, 1.5, -4.0 eV, 1.0e-7 m, 3, 1.0e-7 m^2 s^-1, 1.0e18 cm^-3, 12, -5 eV, 1.0e-7 m, 3, 1.0e-7 m^2 s^-1), 366×21 DataFrame
 Row │ t               V            j                light    nₑ               ⋯
     │ Quantity…       Quantity…    Quantity…        Float64  Array…           ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │  0.0 s          -0.5 V       -223.006 A m^-2      1.0  Unitful.Quantity ⋯
   2 │  2.00076e-6 s   -0.5 V       -223.006 A m^-2      1.0  Unitful.Quantity
   3 │  1.15639e-5 s   -0.499998 V  -223.006 A m^-2      1.0  Unitful.Quantity
   4 │  0.000103741 s  -0.499979 V  -223.006 A m^-2      1.0  Unitful.Quantity
   5 │  0.000567422 s  -0.499887 V  -223.006 A m^-2      1.0  Unitful.Quantity ⋯
   6 │  0.00208718 s   -0.499583 V  -223.006 A m^-2      1.0  Unitful.Quantity
   7 │  0.00579951 s   -0.49884 V   -223.006 A m^-2      1.0  Unitful.Quantity
   8 │  0.0131985 s    -0.49736 V   -223.005 A m^-2      1.0  Unitful.Quantity
  ⋮  │       ⋮              ⋮              ⋮            ⋮                      ⋱
 360 │ 20.966 s        -0.293202 V  -223.057 A m^-2      1.0  Unitful.Quantity ⋯
 361 │ 21.1525 s       -0.330496 V  -223.058 A m^-2      1.0  Unitful.Quantity
 362 │ 21.339 s        -0.367791 V  -223.058 A m^-2      1.0  Unitful.Quantity
 363 │ 21.5254 s       -0.405086 V  -223.058 A m^-2      1.0  Unitful.Quantity
 364 │ 21.7119 s       -0.442381 V  -223.059 A m^-2      1.0  Unitful.Quantity ⋯
 365 │ 21.8984 s       -0.479675 V  -223.059 A m^-2      1.0  Unitful.Quantity
 366 │ 22.0 s          -0.5 V       -223.059 A m^-2      1.0  Unitful.Quantity
                                                 17 columns and 351 rows omitted, [-0.25, -0.2499713357012002, -0.24993947809533124, -0.24990407239148255, -0.24986472460467835, -0.2498209972790036, -0.2497724047565189, -0.24971840794668904, -0.24965840854725224, -0.24959174266352288  …  -0.00048232623090260773, -0.00040825733647711915, -0.00034159145274775815, -0.00028159205331096393, -0.0002275952434810935, -0.0001790027209964129, -0.00013527539532166843, -9.592760851746818e-5, -6.0521904668750515e-5, -2.8664298799802035e-5], [0.0, 2.8655728991078266e-5, 5.8109454289356766e-5, 8.83833492685282e-5, 0.0001195002006493806, 0.00015148342530896253, 0.00018435708754205304, 0.00021814591678570583, 0.0002528753258191907, 0.0002885714294518782  …  0.9997114285705482, 0.9997471246741808, 0.9997818540832143, 0.999815642912458, 0.9998485165746911, 0.9998804997993507, 0.9999116166507315, 0.9999418905457107, 0.999971344271009, 1.0], [1.0000286642987999, 1.0000605219046688, 1.0000959276085175, 1.0001352753953217, 1.0001790027209965, 1.0002275952434811, 1.000281592053311, 1.0003415914527478, 1.0004082573364772, 1.0004823262309026  …  1.2495917426635228, 1.2496584085472522, 1.249718407946689, 1.2497724047565189, 1.2498209972790035, 1.2498647246046783, 1.2499040723914825, 1.2499394780953312, 1.2499713357012001, 1.25])

The ProblemSolution object contains also grid and spatial information. All timesteps are stored in a DataFrame an can be acessed via the df field of sol. If we just want to plot the IV characteristics we can do:

using Plots
using UnitfulRecipes  # To interface Unitful with Plots

sol=sol.df
sol.j = sol.j .|>u"mA/cm^2" # scale to common units
plt = plot(sol.V[sol.fwd], sol.j[sol.fwd],label="Forward", ylims=(-25,40),xlims=(-0.5,1.3),legend=:topleft);
plot!(plt,sol.V[.!sol.fwd], sol.j[.!sol.fwd],label="Backward");
plt

Jsc vs. Voc Curve

The wrapper implements $j_{sc}(V_{oc})$ simulation by two consecutive simulation runs, where the illumination is increased exponentially over time. This reflects the aspect of a slow $V_{oc}$ built up under low illumination intensities.

Open Circuit Voltage Decay (OCVD)

To simulate a open circuit voltage decay a OCVDProblem is implemented:

prob_ocvd = OCVDProblem(
    parm,       # input parameter set
    50u"s",     # illumination time
    1e5u"s",    # time the decay will be simulated to
)

sol = solve(prob_ocvd)
plot(sol.t_decay,sol.V_decay,xscale=:log10,xlims=(1e-8,10))