Predefined Protocols

Convenient wrappers are predefined to simplify the simulation of most important protocols.

IV Simulations

The simulation of an IV curve can be done by the IVProblem constructor

AnnA.IVProblem — Type

IVProblem(
    parm::Parameters,
    range::Union{AbstractArray,Tuple},
    rate;
    double_sweep = true,
    alg_control = AlgControl(dtmin = 1e-20,
        dt = 1e-6,
        reltol = 1e-4,
        abstol = 1e-12,
        tend = abs(sum(diff(range))) / abs(rate) * (1+double_sweep)
    ),  
)

Creates an IVProblem. The voltage parameter V defined in parm will be overwritten by the linear voltage sweep defined as V = t-> first(range) + rate*t. AlgControl.tend is forced to be abs(sum(diff(range))) / abs(rate) * (1+double_sweep), an overwrite can be done on the finalized object using Setfield: p = Setfield.setproperties(p::IVProblem, alg_control=AlgControl(...)).

Units

If no units provided for range and rate, V and V/s is assumed.

Here is a simple example on how to use

using AnnA
using Unitful, UnitfulRecipes
parm = Parameters(light = t -> 1.0,
    vₙₕ= 10u"m/s" ,                 # electron surface recombination vel. at HTM
    vₚₕ= 0.01u"m/s" ,               # hole surface recombination vel. at HTM
    N=500,                          # grid size
    N₀=1e18u"cm^-3"                 # ionic concentration
)
prob = IVProblem(parm, [-0.5,1.7]u"V", 0.2u"V/s")
sol  = solve(prob)

AnnA.ProblemSolution{AnnA.IV}(t: [0.0, 2.000763597981684e-7, 8.221472611440358e-7, 3.9417054233908385e-6, 1.3455051878708356e-5, 3.548841013630263e-5, 7.778996217177347e-5, 0.00014893706934888333, 0.0002618549367347428, 0.00043465002987396026  …  0.10211557976444578, 0.10299994142824813, 0.10388430309205048, 0.10476866475585282, 0.10565302641965517, 0.10653738808345752, 0.10742174974725986, 0.10830611141106221, 0.10919047307486456, 0.10995801813963932]
u: [[0.03281147383686212, 0.03295358939846525, 0.0331001767897792, 0.033251390579330194, 0.03340739143199689, 0.033568346388317925, 0.033734429158468665, 0.03390582043175775, 0.03408270820255419, 0.03426528811361262  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.03281147410628766, 0.03295358952515975, 0.03310017684247251, 0.03325139059897545, 0.03340739143900604, 0.03356834639119424, 0.03373442916021355, 0.03390582043326419, 0.034082708204036874, 0.03426528811510693  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.03281147610949621, 0.03295359099870228, 0.03310017788536584, 0.03325139130775858, 0.033407391902257826, 0.03356834668382696, 0.03373442934066347, 0.03390582054390756, 0.03408270827360225, 0.03426528816185463  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.03281149794318896, 0.03295361006743323, 0.03310019437840539, 0.03325140542986793, 0.0334074038698412, 0.033568356720121456, 0.03373443767050673, 0.033905827388568295, 0.03408271384589311, 0.034265292661769135  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.03281162748440158, 0.0329537312421415, 0.033100307345263404, 0.033251510374983, 0.033407501006785155, 0.033568446289562376, 0.03373451993943221, 0.03390590264908789, 0.03408278241359818, 0.03426535487362971  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.03281214116710493, 0.03295422625084647, 0.03310078364076157, 0.03325196794789227, 0.03340793987946485, 0.03356886651797991, 0.03373492161495583, 0.03390628590017701, 0.034083147407358586, 0.03426570181719685  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.03281367715736086, 0.03295572850916145, 0.033102251748736655, 0.03325340151021171, 0.03340933852646554, 0.03357022990833805, 0.033736249438485175, 0.03390757788073127, 0.03408440330582799, 0.034266921434586074  …  1.0, 1.0, 1.0, 0.9999999999999999, 0.9999999999999999, 1.0, 1.0, 1.0, 1.0, 1.0], [0.032817430532799285, 0.0329594298385118, 0.033105900093459185, 0.033256995940629315, 0.033412878124250055, 0.033573713769181175, 0.03373967667496084, 0.03391094762535542, 0.034087714714321295, 0.03427017368934402  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.032825693628042406, 0.03296761962115877, 0.03311401499456165, 0.0332650343809459, 0.03342083851626739, 0.03358159451933817, 0.03374747618608898, 0.03391866429934853, 0.03409534695504811, 0.03427771990581753  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.03284273556839318, 0.03298456695898566, 0.03313086550334716, 0.033281785802238444, 0.033437488561001864, 0.03359814086935049, 0.03376391649584033, 0.033934996197875614, 0.03411156804815702, 0.03429382777853945  …  0.9999999999999999, 0.9999999999999999, 0.9999999999999999, 0.9999999999999999, 0.9999999999999999, 0.9999999999999999, 1.0, 1.0, 1.0, 1.0]  …  [0.22370018652591447, 0.2243222011549479, 0.2249621548507976, 0.22562057853437886, 0.2262980194327194, 0.22699504159843392, 0.22771222644558609, 0.2284501733023248, 0.22920949998070475, 0.22999084336409104  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.2163292483016836, 0.2169409647821046, 0.21757035873581243, 0.218217955336408, 0.21888429599864298, 0.21956993890207546, 0.2202754595316134, 0.22100145123537165, 0.2217485258002931, 0.22251731404597866  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.20908674020103135, 0.20968799357125567, 0.21030665715479155, 0.21094325016257712, 0.2115983079727633, 0.21227238265815795, 0.212966043531041, 0.2136798777058154, 0.21441449067998458, 0.21517050693394488  …  1.0000000000000002, 1.0, 1.0, 1.0000000000000002, 1.0000000000000002, 1.0000000000000002, 1.0000000000000002, 1.0, 1.0, 1.0], [0.2019704900138632, 0.20256110374501737, 0.2031688544001741, 0.203794255000001, 0.20443783464901494, 0.20510013906640437, 0.20578173113468878, 0.20648319146671812, 0.20720511899154304, 0.20794813155968742  …  0.9999999999999999, 0.9999999999999999, 1.0, 0.9999999999999999, 0.9999999999999999, 0.9999999999999999, 1.0, 1.0, 1.0, 1.0], [0.1949794482795019, 0.19555923881200743, 0.19615588670809506, 0.1967698985612812, 0.1974017969557469, 0.19805212100011005, 0.19872142687948294, 0.19941028842635528, 0.2001192977108731, 0.20084906565108696  …  1.0, 1.0, 1.0, 1.0, 1.0000000000000002, 1.0, 1.0, 1.0, 1.0, 1.0], [0.18811346421572528, 0.18868224503645276, 0.18926759725805384, 0.18987002080357074, 0.19049003148336746, 0.1911281615314053, 0.19178496016023336, 0.19246099413526987, 0.19315684836898467, 0.19387312653559685  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.18137309403947316, 0.18193067934118148, 0.18250454364849983, 0.18309517996690974, 0.18370309707543717, 0.18432882006496737, 0.18497289089568372, 0.1856358689742431, 0.18631833175133525, 0.18702087534028108  …  1.0, 1.0, 1.0000000000000002, 1.0000000000000002, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.17475943648871728, 0.17530564444514743, 0.17586783264097125, 0.17644648691578788, 0.1770421087583157, 0.17765521584626578, 0.17828634260572726, 0.17893604079071243, 0.17960488008354594, 0.18029344871679442  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], [0.1682739915456829, 0.16880864722981564, 0.16935897816110643, 0.16992546276609158, 0.17050859498520504, 0.1711088848137141, 0.17172685886253233, 0.17236306093959217, 0.17301805265249856, 0.17369241403319843  …  1.0000000000000002, 1.0000000000000002, 1.0, 1.0, 1.0000000000000002, 1.0000000000000002, 1.0000000000000002, 1.0, 1.0, 1.0], [0.16275052037534096, 0.16327501524010374, 0.16381491576828647, 0.16437069375474067, 0.1649428363820379, 0.16553184676191393, 0.1661382444968886, 0.16676256626277294, 0.1674053664128149, 0.16806721760425355  …  1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0]], Parameters(500, 8.854187817e-12 F m^-1, 1.6021766209e-19 C, 8.61733035e-5 eV K^-1, 4.0e-7 m, 24.1, -3.7 eV, -5.3 eV, 0 m^-3, 0.00017 m^2 s^-1, 0.00017 m^2 s^-1, 0.2, 0.2, 1.0e18 cm^-3, 6.5e-8 m^2 s^-1, 0.58 eV, false, 300 K, 1.3e7 m^-1, 1.4e21 m^-2 s^-1, 1, Main.__atexample__named__iv.var"#1#2"(), AnnA.var"#73#74"{Vector{Float64}, Float64, Float64, Bool}([-0.5, 1.7], 0.14279966607226333, 2.2, false), 1.0e6 m^2 V A^-1, 0.0001 m^2 V A^-1, 3.0e-7 s, 3.0e-7 s, 3.22e-17 m^3 s^-1, 0 m^4 s^-1, 0 m^4 s^-1, 0 m s^-1, 0 m s^-1, 10 m s^-1, 0.01 m s^-1, 1.0e18 cm^-3, 1.5, -4.0 eV, 1.0e-7 m, 3, 1.0e-7 m^2 s^-1, 1.0e18 cm^-3, 12, -5 eV, 1.0e-7 m, 3, 1.0e-7 m^2 s^-1), 242×22 DataFrame
 Row │ t               V            j                light    nₑ               ⋯
     │ Quantity…       Quantity…    Quantity…        Float64  Array…           ⋯
─────┼──────────────────────────────────────────────────────────────────────────
   1 │  0.0 s          -0.5 V       -223.006 A m^-2      1.0  Unitful.Quantity ⋯
   2 │  4.00305e-5 s   -0.499992 V  -223.006 A m^-2      1.0  Unitful.Quantity
   3 │  0.000164492 s  -0.499967 V  -223.006 A m^-2      1.0  Unitful.Quantity
   4 │  0.000788642 s  -0.499842 V  -223.006 A m^-2      1.0  Unitful.Quantity
   5 │  0.00269204 s   -0.499462 V  -223.006 A m^-2      1.0  Unitful.Quantity ⋯
   6 │  0.00710039 s   -0.49858 V   -223.006 A m^-2      1.0  Unitful.Quantity
   7 │  0.0155639 s    -0.496887 V  -223.005 A m^-2      1.0  Unitful.Quantity
   8 │  0.0297988 s    -0.49404 V   -223.004 A m^-2      1.0  Unitful.Quantity
  ⋮  │       ⋮              ⋮              ⋮            ⋮                      ⋱
 236 │ 20.9617 s       -0.292347 V  -223.058 A m^-2      1.0  Unitful.Quantity ⋯
 237 │ 21.1387 s       -0.327735 V  -223.058 A m^-2      1.0  Unitful.Quantity
 238 │ 21.3156 s       -0.363123 V  -223.059 A m^-2      1.0  Unitful.Quantity
 239 │ 21.4926 s       -0.398511 V  -223.059 A m^-2      1.0  Unitful.Quantity
 240 │ 21.6695 s       -0.433899 V  -223.059 A m^-2      1.0  Unitful.Quantity ⋯
 241 │ 21.8464 s       -0.469286 V  -223.059 A m^-2      1.0  Unitful.Quantity
 242 │ 22.0 s          -0.5 V       -223.06 A m^-2       1.0  Unitful.Quantity
                                                 18 columns and 227 rows omitted, Unitful.Quantity{Float64, 𝐋, Unitful.FreeUnits{(m,), 𝐋, nothing}}[-1.0e-7 m, -9.998853428048007e-8 m, -9.997579123813249e-8 m, -9.996162895659301e-8 m, -9.994588984187134e-8 m, -9.992839891160143e-8 m, -9.990896190260756e-8 m, -9.988736317867561e-8 m, -9.986336341890089e-8 m, -9.983669706540915e-8 m  …  -1.929304923610431e-10 m, -1.6330293459084766e-10 m, -1.3663658109910326e-10 m, -1.1263682132438556e-10 m, -9.10380973924374e-11 m, -7.160108839856516e-11 m, -5.411015812866737e-11 m, -3.8371043406987266e-11 m, -2.4208761867500205e-11 m, -1.1465719519920814e-11 m], Unitful.Quantity{Float64, 𝐋, Unitful.FreeUnits{(m,), 𝐋, nothing}}[0.0 m, 1.1462291596431307e-11 m, 2.3243781715742705e-11 m, 3.535333970741128e-11 m, 4.780008025975224e-11 m, 6.059337012358501e-11 m, 7.374283501682122e-11 m, 8.725836671428233e-11 m, 1.0115013032767627e-10 m, 1.1542857178075127e-10 m  …  3.9988457142821927e-7 m, 3.998988498696723e-7 m, 3.999127416332857e-7 m, 3.9992625716498315e-7 m, 3.999394066298764e-7 m, 3.9995219991974026e-7 m, 3.999646466602926e-7 m, 3.9997675621828425e-7 m, 3.999885377084036e-7 m, 4.0e-7 m], Unitful.Quantity{Float64, 𝐋, Unitful.FreeUnits{(m,), 𝐋, nothing}}[4.0001146571951993e-7 m, 4.000242087618675e-7 m, 4.00038371043407e-7 m, 4.0005411015812863e-7 m, 4.000716010883986e-7 m, 4.000910380973924e-7 m, 4.0011263682132437e-7 m, 4.001366365810991e-7 m, 4.0016330293459087e-7 m, 4.00192930492361e-7 m  …  4.998366970654091e-7 m, 4.998633634189008e-7 m, 4.998873631786755e-7 m, 4.999089619026075e-7 m, 4.999283989116014e-7 m, 4.999458898418713e-7 m, 4.99961628956593e-7 m, 4.999757912381324e-7 m, 4.9998853428048e-7 m, 5.0e-7 m])

The ProblemSolution object contains also grid and spatial information. All timesteps are stored in a DataFrame an can be acessed via the df field of sol. If we just want to plot the IV characteristics we can do:

using Plots
using UnitfulRecipes  # To interface Unitful with Plots
gr()
sol=sol.df
sol.j = sol.j .|>u"mA/cm^2" # scale to common units
plt = plot(sol.V[sol.fwd], sol.j[sol.fwd],label="Forward", ylims=(-25,40),xlims=(-0.5,1.3),legend=:topleft);
plot!(plt,sol.V[.!sol.fwd], sol.j[.!sol.fwd],label="Backward");
plt

Jsc vs. Voc Curve

The wrapper implements $j_{sc}(V_{oc})$ simulation by two consecutive simulation runs, where the illumination is increased exponentially over time. This reflects the aspect of a slow $V_{oc}$ built up under low illumination intensities.

Open Circuit Voltage Decay (OCVD)

To simulate a open circuit voltage decay a OCVDProblem is implemented:

prob_ocvd = OCVDProblem(
    parm,       # input parameter set
    50u"s",     # illumination time
    1e5u"s",    # time the decay will be simulated to
)

sol = solve(prob_ocvd)
plot(sol.t_decay,sol.V_decay,xscale=:log10,xlims=(1e-8,1e5))